Refactor for independent tasks for different bots
