+++ lib/plugins/digging.js
@@ -1,5 +1,5 @@
 const { performance } = require('perf_hooks')
-const { createDoneTask, createCompatibleTask, tasks, waitForTaskCompatibility } = require('../promise_utils')
+const { createDoneTask, tasks } = require('../tasks')
 const BlockFaces = require('prismarine-world').iterators.BlockFace
 const { Vec3 } = require('vec3')
 
@@ -92,8 +92,8 @@ function inject (bot) {
       }
     }
 
-    await waitForTaskCompatibility(tasks.digging_dig)
-    diggingTask = createCompatibleTask(tasks.digging_dig)
+    await bot.waitForTaskCompatibility(tasks.digging_dig)
+    diggingTask = bot.createCompatibleTask(tasks.digging_dig)
     bot._client.write('block_dig', {
       status: 0, // start digging
       location: block.position,
