+++ lib/plugins/fishing.js
@@ -1,5 +1,5 @@
 const { Vec3 } = require('vec3')
-const { createDoneTask, createCompatibleTask, tasks } = require('../promise_utils')
+const { createDoneTask, tasks } = require('../tasks')
 
 module.exports = inject
 
@@ -46,7 +46,7 @@ function inject (bot) {
     }
 
     // await waitForTaskCompatibility(tasks.fishing_fish)
-    fishingTask = createCompatibleTask(tasks.fishing_fish)
+    fishingTask = bot.createCompatibleTask(tasks.fishing_fish)
 
     bot.activateItem()
