+++ lib/plugins/inventory.js
@@ -1,7 +1,8 @@
 const assert = require('assert')
 const { Vec3 } = require('vec3')
 const { once } = require('events')
-const { sleep, createDoneTask, createCompatibleTask, withTimeout, tasks } = require('../promise_utils')
+const { sleep, withTimeout } = require('../promise_utils')
+const { createDoneTask, tasks } = require('../tasks')
 
 module.exports = inject
 
@@ -83,7 +84,7 @@ function inject (bot, { hideErrors }) {
       throw new Error('Food is full')
     }
 
-    eatingTask = createCompatibleTask(tasks.inventory_consume)
+    eatingTask = bot.createCompatibleTask(tasks.inventory_consume)
 
     activateItem()
