+++ lib/plugins/physics.js
@@ -3,7 +3,7 @@ const assert = require('assert')
 const math = require('../math')
 const conv = require('../conversions')
 const { performance } = require('perf_hooks')
-const { createDoneTask, createCompatibleTask, tasks } = require('../promise_utils')
+const { createDoneTask } = require('../tasks')
 
 const { Physics, PlayerState } = require('prismarine-physics')
 
@@ -225,7 +225,7 @@ function inject (bot, { physicsEnabled }) {
     if (!lookingTask.done) {
       lookingTask.cancel() // cancel the previous one
     }
-    lookingTask = createCompatibleTask(tasks.physics_look)
+    lookingTask = bot.createCompatibleTask(tasks.physics_look)
 
     // this is done to bypass certain anticheat checks that detect the player's sensitivity
     // by calculating the gcd of how much they move the mouse each tick
