+++ lib/promise_utils.js
@@ -1,100 +1,7 @@
-const { on } = require('events')
-const EventEmitter = require('events')
-const TaskData = require('./task_data.json')
-/** @type {Task[]} */
-const CurrentTasks = []
-const taskUpdate = new EventEmitter()
+const { createBlankTask } = require('./tasks')
 
 function sleep (ms) {
   return new Promise(resolve => setTimeout(resolve, ms))
-}
-
-/**
- * @param {Task} task
- */
-function removeTask (task) {
-  const i = CurrentTasks.indexOf(task)
-  if (i < 0) return
-  CurrentTasks.splice(i, 1)
-  if (task.done === false) {
-    task.cancel()
-  }
-  taskUpdate.emit('update')
-}
-
-function taskCompatibleWithRunningTask (taskIdentifier) {
-  // Only one task of a compatible task group can be active at any given time.
-  for (const group of TaskData.compatibility) {
-    // If not part of group skip to next group
-    if (!group.includes(taskIdentifier)) continue
-    // If taskIdentifier is part of this group and another running task is part of the same group throw error
-    for (const task of CurrentTasks) {
-      if (group.find(gi => gi === task.taskIdentifier) && group.includes(taskIdentifier)) {
-        return { status: false, task: task }
-      }
-    }
-  }
-  return { status: true }
-}
-
-/**
- * @property {string?} taskIdentifier The identifier name of the task
- * @property {Promise<void>} promise A promise that is resolved when the task is done or rejects if the task is canceled
- * @property {boolean} done If the task is done or not
- */
-class Task {
-  constructor (name) {
-    this.taskIdentifier = name
-    this.promise = new Promise((resolve, reject) => {
-      this._resolver = resolve
-      this._rejecter = reject
-    })
-    this.done = false
-  }
-
-  finish (...data) {
-    this.done = true
-    removeTask(this)
-    this._resolver(...data)
-  }
-
-  cancel (...data) {
-    this.done = true
-    removeTask(this)
-    this._rejecter(...data)
-  }
-}
-
-function createCompatibleTask (taskIdentifier) {
-  // console.info('Current tasks:', CurrentTasks, 'New task', taskIdentifier)
-  if (!taskIdentifier || !Object.values(TaskData.tasks).includes(taskIdentifier)) throw new Error('Invalid task identifier ' + String(taskIdentifier))
-  if (CurrentTasks.find(t => t.taskIdentifier === taskIdentifier)) throw new Error(`Task ${taskIdentifier} already running`)
-  // Check task compatibility
-  const { status, task: incompatiblyTask } = taskCompatibleWithRunningTask(taskIdentifier)
-  if (status === false) {
-    // console.info(CurrentTasks, status)
-    throw new Error(`Task ${taskIdentifier} incompatible with already running task ${incompatiblyTask.taskIdentifier}`)
-  }
-  const task = new Task(taskIdentifier)
-  CurrentTasks.push(task)
-  return task
-}
-
-async function waitForTaskCompatibility (taskIdentifier) {
-  for await (const _ of on(taskUpdate, 'update')) { // eslint-disable-line no-unused-vars
-    if (!taskCompatibleWithRunningTask(taskIdentifier)) continue
-    return
-  }
-}
-
-function createBlankTask () {
-  return new Task(undefined)
-}
-
-function createDoneTask () {
-  const task = createBlankTask()
-  task.finish()
-  return task
 }
 
 /**
@@ -147,12 +54,6 @@ function withTimeout (promise, timeout) {
 
 module.exports = {
   sleep,
-  createBlankTask,
-  createCompatibleTask,
-  waitForTaskCompatibility,
-  createDoneTask,
-  taskCompatibleWithRunningTask,
-  tasks: TaskData.tasks,
   onceWithCleanup,
   withTimeout
 }
